+++ lib/temperature/ThermistorConstants.h
@@ -2,7 +2,7 @@
 
 // Beta coefficient (β)
 namespace ThermistorConstants {
-    constexpr float BETA_3435 = 3435.0f;
-    constexpr float BETA_3950 = 3950.0f;
-    constexpr float BETA_4250 = 4250.0f;
+    constexpr uint16_t BETA_3435 = 3435;
+    constexpr uint16_t BETA_3950 = 3950;
+    constexpr uint16_t BETA_4250 = 4250;
 }
